Adding filters to the "Charts" Page
Adding buttons to the "Charts" page so that users can select subsets of
students on which to get statistics. Note that the better implementation
of this would be to dynamically fetch the charts using AJAX, but I don't
know how to do that. Hopefully the refreshes won't be too annoying :/